fix(quality): clear the five staged Sonar new-code findings
java:S1845 (BLOCKER) in ForwardedHeaderResolver: the ForwardedResult
constant UNRESOLVABLE and the record accessor unresolvable() differed
only by case. Rename the constant to UNRESOLVABLE_RESULT and update the
enclosing class comment's {@link} reference. Clean break, no shim -
both the constant and the record are private to a nested type and every
caller is in this file.

java:S1905 in RetryConfig line 197: drop the redundant (double) cast on
maxDelay.toMillis() in the Math.min call, since exponentialDelay is
already double and the argument widens implicitly. The Math.pow cast on
line 192 is a separate finding and is left untouched.

java:S127 in LengthValidationStage.utf8ByteLength: the loop counter was
advanced from inside the for body to skip a low surrogate. Convert to a
while loop that advances by a computed step, so the counter moves in one
place. The surrogate-pair 4-byte count is unchanged.

java:S127 in DecodingStage.validateDecodedCharacters: the for loop had
an empty update clause and advanced by Character.charCount(cp) in the
body. Convert to a while loop; code-point iteration semantics are
preserved exactly.

java:S5778 in HeaderAndContentTypeEnforcementRegressionTest line 228:
the assertThrows lambda held two invocations. Hoist pipeline() into a
local so only the throwing call remains inside the lambda.

Each file's diff touches only its one named finding.

+++ cui-http-core/src/main/java/de/cuioss/http/client/adapter/RetryConfig.java
@@ -194,7 +194,7 @@ boolean idempotentOnly
         // Cap the base exponential delay at maxDelay BEFORE applying jitter, so that once the
         // exponential growth reaches the cap the retries still spread out (± jitter around the cap)
         // instead of all firing in lockstep at exactly maxDelay (the thundering herd jitter prevents).
-        double cappedDelay = Math.min(exponentialDelay, (double) maxDelay.toMillis());
+        double cappedDelay = Math.min(exponentialDelay, maxDelay.toMillis());
 
         // Apply jitter: delay * (1 ± jitter)
         // Random value between -1.0 and 1.0
